Pull non-kivy image utils out of image_utils.py
